product-service/requirements.txt
annotated-types==0.6.0
anyio==4.0.0
atomicwrites==1.4.1
attrs==23.2.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
coverage==7.5.1
crypto==1.4.1
cryptography==42.0.5
dnspython==2.5.0
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.1
filelock==3.13.4
firebase-admin==6.4.0
flake8==4.0.1
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
ifaddr==0.2.0
iniconfig==2.0.0
jose==1.0.0
markdown-it-py==3.0.0
mccabe==0.6.1
mdurl==0.1.2
motor==3.1.1
msgpack==1.0.8
Naked==0.1.32
odmantic==1.0.1
packaging==24.0
passlib==1.7.4
pillow==10.3.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
py==1.11.0
py_eureka_client==0.11.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.8.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyflakes==2.4.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.2
pyparsing==3.1.2
pytest==8.0.0
pytest-asyncio==0.20.0
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-mock-resources==2.10.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
python-on-whales==0.70.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.29
starlette==0.37.2
tomli==2.0.1
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.19.0
